// src/app/api/model/user.ts
/**
 * GameHubAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface User {
  id: string;
  username: string;
  displayName: string;
  email: string;
  hash: string;
  salt: string;
  bio?: string;
  profilePictureURL?: string;
  bannerURL?: string;
  creationDate: Date;
}